refactor: split commands into modules
