+++ src/commands/clean.rs
@@ -1,0 +1,54 @@
+use crate::{
+    output::status,
+    project::{
+        ProjectDiscoveryError, cache_dir_path, find_project_root, project_library_root_path,
+    },
+};
+use miette::Diagnostic;
+use std::{fs, path::Path};
+use thiserror::Error;
+
+#[derive(Debug, Error, Diagnostic)]
+pub(crate) enum Error {
+    #[error(transparent)]
+    #[diagnostic(transparent)]
+    ProjectDiscovery(#[from] ProjectDiscoveryError),
+
+    #[error("failed to remove {label} at {path}")]
+    #[diagnostic(code(rpx::clean::remove_failed))]
+    RemoveFailed {
+        label: String,
+        path: String,
+        #[source]
+        source: std::io::Error,
+    },
+}
+
+pub(crate) fn run() -> Result<(), Error> {
+    let current_dir = find_project_root()?;
+    let mut removed_any = false;
+
+    removed_any |=
+        remove_dir_if_exists(&project_library_root_path(&current_dir), "project library")?;
+    removed_any |= remove_dir_if_exists(&cache_dir_path(), "cache directory")?;
+
+    if removed_any {
+        status("Removed project library and cache directories");
+    } else {
+        status("Project library and cache directories are already clean");
+    }
+    Ok(())
+}
+
+fn remove_dir_if_exists(path: &Path, label: &str) -> Result<bool, Error> {
+    if !path.exists() {
+        return Ok(false);
+    }
+
+    fs::remove_dir_all(path).map_err(|source| Error::RemoveFailed {
+        label: label.to_string(),
+        path: path.display().to_string(),
+        source,
+    })?;
+    Ok(true)
+}
