+++ src/commands/mod.rs
@@ -1,2 +1,9 @@
+pub(crate) mod add;
+pub(crate) mod clean;
 pub(crate) mod init;
+pub(crate) mod lock;
+pub(crate) mod remove;
 pub(crate) mod repo;
+pub(crate) mod run;
+pub(crate) mod status;
+pub(crate) mod sync;
