+++ src/commands/run.rs
@@ -1,0 +1,48 @@
+use crate::{
+    project::{ProjectDiscoveryError, find_project_root, project_library_path},
+    r::RVirtualEnv,
+};
+use miette::Diagnostic;
+use thiserror::Error;
+use tokio::process::Command;
+
+#[derive(Debug, Error, Diagnostic)]
+pub(crate) enum Error {
+    #[error(transparent)]
+    #[diagnostic(transparent)]
+    ProjectDiscovery(#[from] ProjectDiscoveryError),
+
+    #[error("failed to run {program}")]
+    #[diagnostic(code(rpx::run::command_failed))]
+    CommandFailed {
+        program: String,
+        #[source]
+        source: std::io::Error,
+    },
+}
+
+pub(crate) async fn run(command: &[String]) -> Result<(), Error> {
+    let project_path = find_project_root()?;
+    let (program, args) = command
+        .split_first()
+        .expect("run command requires at least one argument");
+    let project_library = project_library_path(&project_path);
+
+    let status = Command::with_venv(program, &project_library)
+        .args(args)
+        .status()
+        .await
+        .map_err(|source| Error::CommandFailed {
+            program: program.clone(),
+            source,
+        })?;
+
+    exit_with_status(status.code());
+    Ok(())
+}
+
+fn exit_with_status(code: Option<i32>) {
+    if code != Some(0) {
+        std::process::exit(code.unwrap_or(1));
+    }
+}
